feat!: moved bridges into *generic*, with new bridges object to hold the base ones
